Fixed Merge Conflict and Reviewed Code
Made changes from code review as well as merged latest MASTER to HEAD.
