+++ Assets/Scripts/Controllers/CharController.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Managers;
 using UnityEngine;
 using Debug = UnityEngine.Debug;
