--- Assets/Scripts/Managers/OverlayManager.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 04a33d40a44e4596a54c4c9a3b9b6a00
-timeCreated: 1624465514
