+++ Assets/prefabs/GameMenu.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: e737f6ad07b1f4e27ac24c08a3d4dc21
-DefaultImporter:
+guid: b2249f5e8ed174b0e8f4f1f46b90d0ea
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
